requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.5.0=pypi_0
anyio=3.7.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
cryptography=41.0.4=pypi_0
dataclasses-json=0.6.0=pypi_0
dnspython=2.4.2=pypi_0
email-validator=2.0.0.post2=pypi_0
exceptiongroup=1.1.3=pypi_0
fastapi=0.103.1=pypi_0
frozenlist=1.4.0=pypi_0
greenlet=2.0.2=pypi_0
h11=0.14.0=pypi_0
httpcore=0.18.0=pypi_0
httptools=0.6.0=pypi_0
httpx=0.25.0=pypi_0
idna=3.4=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
langchain=0.0.300=pypi_0
langsmith=0.0.40=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lxml=4.9.3=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
numexpr=2.8.6=pypi_0
numpy=1.26.0=pypi_0
openai=0.28.0=pypi_0
openssl=3.0.11=h7f8727e_2
orjson=3.9.7=pypi_0
packaging=23.1=pypi_0
pdfminer-six=20221105=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=py39h06a4308_0
pycparser=2.21=pypi_0
pydantic=2.3.0=pypi_0
pydantic-core=2.6.3=pypi_0
pydantic-extra-types=2.1.0=pypi_0
pydantic-settings=2.0.3=pypi_0
python=3.9.18=h955ad1f_0
python-docx=0.8.11=pypi_0
python-dotenv=1.0.0=pypi_0
python-multipart=0.0.6=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
setuptools=68.0.0=py39h06a4308_0
sniffio=1.3.0=pypi_0
sqlalchemy=2.0.21=pypi_0
sqlite=3.41.2=h5eee18b_0
starlette=0.27.0=pypi_0
tenacity=8.2.3=pypi_0
tiktoken=0.5.1=pypi_0
tk=8.6.12=h1ccaba5_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023c=h04d1e81_0
ujson=5.8.0=pypi_0
urllib3=2.0.5=pypi_0
uvicorn=0.23.2=pypi_0
uvloop=0.17.0=pypi_0
watchfiles=0.20.0=pypi_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
